SW-2919 Refactor nursery withdrawals service
- no changes except moved all nursery withdrawals api into a single service
